requirements.txt
cachetools==5.3.1
click==8.1.3
Flask==2.2.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.2
google-auth==2.23.0
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
SQLAlchemy==1.4.46
urllib3==1.26.16
Werkzeug==2.2.2
